review: renaming "prtModifier" directory to "modifiers"
