+++ src/serlio/modifiers/polyModifier/polyModifierNode.cpp
