+++ src/serlio/modifiers/polyModifier/polyModifierNode.h
