+++ src/test/tests.cpp
@@ -18,7 +18,9 @@
  */
 
 #include "PRTContext.h"
-#include "prtModifier/RuleAttributes.h"
+
+#include "modifiers/RuleAttributes.h"
+
 #include "util/LogHandler.h"
 #include "util/Utilities.h"
